+++ js/screens/application.js
@@ -1,6 +1,7 @@
 import WelcomeScreen from './welcome';
 import ResultScreen from './result';
 import store from './../data/game-store';
+import switchScreen from './../game/switch-screen';
 
 
 export default class Application {
@@ -11,8 +12,7 @@ export default class Application {
   }
 
   static showGame() {
-    const welcome = new WelcomeScreen(store);
-    welcome.showNextScreen(store);
+    switchScreen();
   }
 
   static showStats() {
