+++ js/view/artist-view.js
@@ -72,7 +72,7 @@ export default class ArtistView extends AbstractView {
 
 
   onAnswerClick() {
-	
+
   }
 
 }
